Composite maximum likelihood estimation for Poisson log-normal model

fit_poilog uses composite maximum likelihood estimation to fit the mean and the covariance of a Poisson log-normal mixed model.

fit_poilog(Y, ncores = 1)

Arguments

Y

Matrix with proteins counts with dimensions (number of cells) x (number of markers)

ncores

Number of CPU cores to use

Value

A list containing

beta

estimated mean parameter

Sigma

estimated covariance parameter

Examples

set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
#> tibble [80 × 7] (S3: tbl_df/tbl/data.frame)
#>  $ donor    : chr [1:80] "pid01" "pid01" "pid01" "pid01" ...
#>  $ condition: Factor w/ 2 levels "control","treatment": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ m01      : num [1:80] 80 6 6 3 17 20 14 90 79 46 ...
#>  $ m02      : num [1:80] 21 3 4 1 1 9 40 17 24 0 ...
#>  $ m03      : num [1:80] 6 2 12 49 3 8 14 6 0 4 ...
#>  $ m04      : num [1:80] 4 0 11 10 0 2 22 8 14 13 ...
#>  $ m05      : num [1:80] 12 0 3 0 0 1 1 4 1 1 ...
Y = dplyr::select(df, names(df)[3:ncol(df)])
fit_poilog(Y)
#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!
#> initial  value 478.229273 
#> iter   2 value 478.170318
#> iter   3 value 478.023225
#> iter   4 value 478.016792
#> iter   5 value 478.008410
#> iter   6 value 477.957787
#> iter   7 value 477.952003
#> iter   8 value 477.931095
#> iter   9 value 477.926304
#> iter  10 value 477.926188
#> iter  11 value 477.926181
#> iter  11 value 477.926181
#> iter  11 value 477.926181
#> final  value 477.926181 
#> converged
#> initial  value 470.263143 
#> iter   2 value 470.221363
#> iter   3 value 470.057484
#> iter   4 value 470.057378
#> iter   5 value 470.031257
#> iter   6 value 470.000750
#> iter   7 value 469.998549
#> iter   8 value 469.994081
#> iter   9 value 469.990876
#> iter   9 value 469.990873
#> iter   9 value 469.990873
#> final  value 469.990873 
#> converged
#> initial  value 511.055697 
#> iter   2 value 511.002016
#> iter   3 value 510.842522
#> iter   4 value 510.826406
#> iter   5 value 510.820218
#> iter   6 value 510.819092
#> iter   7 value 510.813492
#> iter   8 value 510.783336
#> iter   9 value 510.759898
#> iter  10 value 510.759869
#> iter  10 value 510.759869
#> iter  10 value 510.759869
#> final  value 510.759869 
#> converged
#> initial  value 495.615815 
#> iter   2 value 495.328333
#> iter   3 value 494.699716
#> iter   4 value 494.578241
#> iter   5 value 494.552350
#> iter   6 value 494.532772
#> iter   7 value 494.519170
#> iter   8 value 494.504118
#> iter   9 value 494.503657
#> iter  10 value 494.502593
#> iter  11 value 494.502572
#> iter  11 value 494.502572
#> iter  11 value 494.502572
#> final  value 494.502572 
#> converged
#> initial  value 418.356513 
#> iter   2 value 418.315732
#> iter   3 value 418.312660
#> iter   4 value 418.311267
#> iter   5 value 418.304659
#> iter   6 value 418.298701
#> iter   7 value 418.297233
#> iter   8 value 418.296036
#> iter   8 value 418.296036
#> iter   8 value 418.296036
#> final  value 418.296036 
#> converged
#> initial  value 465.339184 
#> iter   2 value 465.307352
#> iter   3 value 465.304627
#> iter   4 value 465.288710
#> iter   5 value 465.265656
#> iter   6 value 465.264605
#> iter   7 value 465.259671
#> iter   8 value 465.259055
#> iter   9 value 465.258701
#> iter   9 value 465.258701
#> iter   9 value 465.258701
#> final  value 465.258701 
#> converged
#> initial  value 449.431843 
#> iter   2 value 448.789899
#> iter   3 value 448.714048
#> iter   4 value 448.704440
#> iter   5 value 448.689689
#> iter   6 value 448.644860
#> iter   7 value 448.641171
#> iter   8 value 448.639129
#> iter   9 value 448.639114
#> iter   9 value 448.639113
#> iter   9 value 448.639113
#> final  value 448.639113 
#> converged
#> initial  value 456.221425 
#> iter   2 value 456.215132
#> iter   3 value 456.213994
#> iter   4 value 456.204136
#> iter   5 value 456.196004
#> iter   6 value 456.195997
#> iter   7 value 456.195966
#> iter   8 value 456.195897
#> iter   9 value 456.195595
#> iter  10 value 456.195581
#> iter  10 value 456.195581
#> iter  10 value 456.195581
#> final  value 456.195581 
#> converged
#> initial  value 440.279043 
#> iter   2 value 439.704799
#> iter   3 value 439.678126
#> iter   4 value 439.663518
#> iter   5 value 439.641024
#> iter   6 value 439.594442
#> iter   7 value 439.588771
#> iter   8 value 439.582973
#> iter   9 value 439.582556
#> iter   9 value 439.582555
#> iter   9 value 439.582555
#> final  value 439.582555 
#> converged
#> initial  value 478.719981 
#> iter   2 value 478.178005
#> iter   3 value 478.047935
#> iter   4 value 478.029376
#> iter   5 value 478.020093
#> iter   6 value 477.962816
#> iter   7 value 477.928805
#> iter   8 value 477.912272
#> iter   9 value 477.912248
#> iter   9 value 477.912248
#> iter   9 value 477.912248
#> final  value 477.912248 
#> converged
#> $beta
#> [1] 1.824467 1.588871 1.466816 2.012334 1.250391
#> 
#> $Sigma
#>             [,1]        [,2]        [,3]        [,4]       [,5]
#> [1,]  2.99162786  0.37749492  0.04242404 -0.01698820 -0.1936195
#> [2,]  0.37749492  1.46707818 -0.67298300  0.08148137  0.3936316
#> [3,]  0.04242404 -0.67298300  1.46907946 -0.13332576 -0.4458472
#> [4,] -0.01698820  0.08148137 -0.13332576  1.43106079  0.8103054
#> [5,] -0.19361945  0.39363160 -0.44584720  0.81030543  4.1613055
#>