HMC Sampling for Logistic Linear Mixed Model

glmm uses Hamiltonion Monte Carlo to sample form an extend logistic linear mixed model. We predict the experimental condition from protein marker expressions.

glmm(
  df_samples_subset,
  protein_names,
  condition,
  group,
  iter = 325,
  warmup = 200,
  num_chains = 1,
  eta = 1
)

Arguments

df_samples_subset
Data frame or tibble with proteins counts, cell condition, and group information

protein_names
A vector of column names of protein to use in the analysis

condition
The column name of the condition variable

group
The column name of the group variable

iter
Number of iteration per chain for the HMC sampler

warmup
Number of warm up steps per chain for the HMC sampler

num_chains
Number of HMC chains to run in parallel

eta
Hyperparametr for LKJ prior

Value

A list of class cytoeffect containing

fit_mcmc

rstan object

protein_names

input protein names

conditions

input condition variable

X

prediction design matrix