Posterior predictive of Poisson log-normal mixed model

Posterior predictive of Poisson log-normal mixed model

posterior_predictive_log_lambda(obj, k = 1, show_donors = TRUE)

Arguments

obj
Object of class cytoeffect_poisson computed using poisson_lognormal

k
Draw from HMC chain

show_donors
Include donor random effect

Value

tibble object

Examples

set.seed(1)
df = simulate_data(n_cells = 10)
str(df)

#> tibble [80 × 7] (S3: tbl_df/tbl/data.frame)
#>  $ donor    : chr [1:80] "pid01" "pid01" "pid01" "pid01" ...
#>  $ condition: Factor w/ 2 levels "control","treatment": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ m01      : num [1:80] 66 3 3 3 7 7 171 46 50 46 ...
#>  $ m02      : num [1:80] 27 3 3 1 5 4 3 20 27 7 ...
#>  $ m03      : num [1:80] 3 6 10 35 1 6 3 0 7 7 ...
#>  $ m04      : num [1:80] 6 0 1 14 1 6 13 4 14 7 ...
#>  $ m05      : num [1:80] 6 1 1 0 5 0 2 8 0 0 ...

fit = poisson_lognormal(df,
                        protein_names = names(df)[3:ncol(df)],
                        condition = "condition",
                        group = "donor",
                        r_donor = 2,
                        warmup = 200, iter = 325,
                        num_chains = 1)

#> 
#> SAMPLING FOR MODEL 'poisson' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000366 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.66 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 325 [  0%]  (Warmup)
#> Chain 1: Iteration:  32 / 325 [  9%]  (Warmup)
#> Chain 1: Iteration:  64 / 325 [ 19%]  (Warmup)
#> Chain 1: Iteration:  96 / 325 [ 29%]  (Warmup)
#> Chain 1: Iteration: 128 / 325 [ 39%]  (Warmup)
#> Chain 1: Iteration: 160 / 325 [ 49%]  (Warmup)
#> Chain 1: Iteration: 192 / 325 [ 59%]  (Warmup)
#> Chain 1: Iteration: 201 / 325 [ 61%]  (Sampling)
#> Chain 1: Iteration: 232 / 325 [ 71%]  (Sampling)
#> Chain 1: Iteration: 264 / 325 [ 81%]  (Sampling)
#> Chain 1: Iteration: 296 / 325 [ 91%]  (Sampling)
#> Chain 1: Iteration: 325 / 325 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 9.29806 seconds (Warm-up)
#> Chain 1:                6.66502 seconds (Sampling)
#> Chain 1:                15.9631 seconds (Total)
#> Chain 1: 

#> Warning: There were 4 divergent transitions after warmup. Increasing adapt_delta above 0.8 may help. See
#> http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

posterior_predictive_log_lambda(fit, k = 1)

#> # A tibble: 80 x 8
#>       m01   m02    m03   m04    m05 condition donor     k
#>     <dbl> <dbl>  <dbl> <dbl>  <dbl> <fct>     <chr> <dbl>
#>  1 0.372  0.943  1.49  0.908  1.72  control   pid01     1
#>  2 3.53   0.861  0     0      0.977 control   pid01     1
#>  3 0      3.57   0.579 2.69  -0.173 control   pid01     1
#>  4 1.86   2.87   1.32  3.39  -0.284 control   pid01     1
#>  5 0      3.28   2.60  2.68   1.33  control   pid01     1
#>  6 0.881  2.28   0     3.34  -0.257 control   pid01     1
#>  7 3.72   2.40   2.50  2.31   0.684 control   pid01     1
#>  8 1.53   2.10   1.25  1.81   1.10  control   pid01     1
#>  9 0.0344 2.02  -0.174 0.653  0.867 control   pid01     1
#> 10 0      2.26   2.65  2.40   0.702 control   pid01     1
#> # … with 70 more rows