Composite maximum likelihood estimation for Poisson log-normal model

fit_poilog uses composite maximum likelihood estimation to fit the mean and the covariance of a Poisson log-normal mixed model.

fit_poilog(Y, ncores = 1)

Arguments

Y
Matrix with proteins counts with dimensions (number of cells) x (number of markers)

ncores
Number of CPU cores to use

Value

A list containing

beta

estimated mean parameter

Sigma

estimated covariance parameter

Examples

set.seed(1)
df = simulate_data(n_cells = 10)
str(df)

#> tibble [80 × 7] (S3: tbl_df/tbl/data.frame)
#>  $ donor    : chr [1:80] "pid01" "pid01" "pid01" "pid01" ...
#>  $ condition: Factor w/ 2 levels "control","treatment": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ m01      : num [1:80] 66 3 3 3 7 7 171 46 50 46 ...
#>  $ m02      : num [1:80] 27 3 3 1 5 4 3 20 27 7 ...
#>  $ m03      : num [1:80] 3 6 10 35 1 6 3 0 7 7 ...
#>  $ m04      : num [1:80] 6 0 1 14 1 6 13 4 14 7 ...
#>  $ m05      : num [1:80] 6 1 1 0 5 0 2 8 0 0 ...

Y = dplyr::select(df, names(df)[3:ncol(df)])
fit_poilog(Y)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!
#> initial  value 488.218706 
#> iter   2 value 488.202446
#> iter   3 value 488.184123
#> iter   4 value 488.178458
#> iter   5 value 488.174502
#> iter   6 value 488.174191
#> iter   7 value 488.173541
#> iter   8 value 488.172719
#> iter   9 value 488.172304
#> iter  10 value 488.170698
#> iter  11 value 488.170597
#> iter  11 value 488.170591
#> iter  11 value 488.170591
#> final  value 488.170591 
#> converged
#> initial  value 473.676894 
#> iter   2 value 473.663855
#> iter   3 value 473.654709
#> iter   4 value 473.654226
#> iter   5 value 473.653907
#> iter   6 value 473.651285
#> iter   7 value 473.650966
#> iter   8 value 473.649877
#> iter   9 value 473.649867
#> iter   9 value 473.649867
#> iter   9 value 473.649867
#> final  value 473.649867 
#> converged
#> initial  value 499.247950 
#> iter   2 value 499.236708
#> iter   3 value 499.227971
#> iter   4 value 499.227467
#> iter   5 value 499.227351
#> iter   6 value 499.225167
#> iter   7 value 499.224534
#> iter   8 value 499.223204
#> iter   8 value 499.223204
#> iter   8 value 499.223204
#> final  value 499.223204 
#> converged
#> initial  value 514.388731 
#> iter   2 value 514.325150
#> iter   3 value 514.298228
#> iter   4 value 514.260266
#> iter   5 value 514.256128
#> iter   6 value 514.255941
#> iter   7 value 514.255809
#> iter   8 value 514.255484
#> iter   8 value 514.255480
#> iter   8 value 514.255480
#> final  value 514.255480 
#> converged
#> initial  value 442.436572 
#> iter   2 value 442.422144
#> iter   3 value 442.418174
#> iter   4 value 442.417839
#> iter   5 value 442.417785
#> iter   6 value 442.417698
#> iter   7 value 442.417617
#> iter   8 value 442.417051
#> iter   9 value 442.417007
#> iter   9 value 442.417006
#> iter   9 value 442.417006
#> final  value 442.417006 
#> converged
#> initial  value 471.729780 
#> iter   2 value 471.704714
#> iter   3 value 471.699789
#> iter   4 value 471.699485
#> iter   5 value 471.699328
#> iter   6 value 471.692210
#> iter   7 value 471.691278
#> iter   8 value 471.689474
#> iter   9 value 471.689333
#> iter   9 value 471.689329
#> iter   9 value 471.689329
#> final  value 471.689329 
#> converged
#> initial  value 487.852001 
#> iter   2 value 487.810875
#> iter   3 value 487.718270
#> iter   4 value 487.714217
#> iter   5 value 487.710942
#> iter   6 value 487.707822
#> iter   7 value 487.706974
#> iter   8 value 487.699798
#> iter   9 value 487.688595
#> iter  10 value 487.688374
#> iter  10 value 487.688374
#> iter  10 value 487.688374
#> final  value 487.688374 
#> converged
#> initial  value 459.151407 
#> iter   2 value 459.143375
#> iter   3 value 459.142864
#> iter   4 value 459.141240
#> iter   5 value 459.136647
#> iter   6 value 459.135234
#> iter   7 value 459.135090
#> iter   8 value 459.132631
#> iter   9 value 459.132426
#> iter   9 value 459.132426
#> iter   9 value 459.132426
#> final  value 459.132426 
#> converged
#> initial  value 471.327148 
#> iter   2 value 471.277941
#> iter   3 value 471.276326
#> iter   4 value 471.275768
#> iter   5 value 471.275423
#> iter   6 value 471.273901
#> iter   7 value 471.273517
#> iter   8 value 471.273017
#> iter   8 value 471.273015
#> iter   8 value 471.273015
#> final  value 471.273015 
#> converged
#> initial  value 498.361960 
#> iter   2 value 498.294869
#> iter   3 value 498.290226
#> iter   4 value 498.289651
#> iter   5 value 498.289583
#> iter   6 value 498.289194
#> iter   7 value 498.289042
#> iter   8 value 498.288718
#> iter   8 value 498.288712
#> iter   8 value 498.288712
#> final  value 498.288712 
#> converged

#> $beta
#> [1] 1.846455 1.602441 1.615194 1.975324 1.703377
#> 
#> $Sigma
#>             [,1]       [,2]       [,3]        [,4]       [,5]
#> [1,]  2.50451452  0.2250056  0.4198440  0.07623263 -0.2668343
#> [2,]  0.22500559  2.0624746 -0.7966710  0.36216416  0.4162924
#> [3,]  0.41984397 -0.7966710  1.4180623 -0.22253076 -0.7418978
#> [4,]  0.07623263  0.3621642 -0.2225308  1.29234194  0.3661240
#> [5,] -0.26683434  0.4162924 -0.7418978  0.36612400  3.3373933
#>