HMC sampling for Poisson log-normal mixed model

poisson_lognormal uses Hamiltonion Monte Carlo to sample form an extend Poisson log-normal mixed model. Each cell and protein marker has its own rate parameter following a linear model.

poisson_lognormal(
  df_samples_subset,
  protein_names,
  condition,
  group,
  r_donor,
  eta = 1,
  iter = 325,
  warmup = 200,
  num_chains = 1,
  adapt_delta = 0.8,
  seed = 1
)

Arguments

df_samples_subset
Data frame or tibble with proteins counts, cell condition, and group information

protein_names
A vector of column names of protein to use in the analysis

condition
The column name of the condition variable

group
The column name of the group variable

r_donor
Rank of the donor random effect covariance matrix

eta
Hyperparameter for LKJ prior

iter
Number of iteration per chain for the HMC sampler

warmup
Number of warm up steps per chain for the HMC sampler

num_chains
Number of HMC chains to run in parallel

adapt_delta
Parameter to control step size of numerical solver

seed
Set seed for HMC sampler (higher value means smaller step size, max is 1)

Value

A list of class cytoeffect_poisson containing

fit_mcmc

rstan object

protein_names

input protein names

condition

input condition variable

group

input group names

df_samples_subset

input df_samples_subset table

Examples

set.seed(1)
df = simulate_data(n_cells = 10)
str(df)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	80 obs. of  7 variables:
#>  $ donor    : chr  "pid01" "pid01" "pid01" "pid01" ...
#>  $ condition: Factor w/ 2 levels "control","treatment": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ m01      : num  66 3 3 3 7 7 171 46 50 46 ...
#>  $ m02      : num  27 3 3 1 5 4 3 20 27 7 ...
#>  $ m03      : num  3 6 10 35 1 6 3 0 7 7 ...
#>  $ m04      : num  6 0 1 14 1 6 13 4 14 7 ...
#>  $ m05      : num  6 1 1 0 5 0 2 8 0 0 ...

fit = poisson_lognormal(df,
                        protein_names = names(df)[3:ncol(df)],
                        condition = "condition",
                        group = "donor",
                        r_donor = 2,
                        warmup = 200, iter = 325, adapt_delta = 0.95,
                        num_chains = 1)

#> 
#> SAMPLING FOR MODEL 'poisson' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00013 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.3 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 325 [  0%]  (Warmup)
#> Chain 1: Iteration:  32 / 325 [  9%]  (Warmup)
#> Chain 1: Iteration:  64 / 325 [ 19%]  (Warmup)
#> Chain 1: Iteration:  96 / 325 [ 29%]  (Warmup)
#> Chain 1: Iteration: 128 / 325 [ 39%]  (Warmup)
#> Chain 1: Iteration: 160 / 325 [ 49%]  (Warmup)
#> Chain 1: Iteration: 192 / 325 [ 59%]  (Warmup)
#> Chain 1: Iteration: 201 / 325 [ 61%]  (Sampling)
#> Chain 1: Iteration: 232 / 325 [ 71%]  (Sampling)
#> Chain 1: Iteration: 264 / 325 [ 81%]  (Sampling)
#> Chain 1: Iteration: 296 / 325 [ 91%]  (Sampling)
#> Chain 1: Iteration: 325 / 325 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 16.1865 seconds (Warm-up)
#> Chain 1:                14.3057 seconds (Sampling)
#> Chain 1:                30.4922 seconds (Total)
#> Chain 1: 

#> Warning: There were 125 transitions after warmup that exceeded the maximum treedepth. Increase max_treedepth above 10. See
#> http://mc-stan.org/misc/warnings.html#maximum-treedepth-exceeded

#> Warning: Examine the pairs() plot to diagnose sampling problems

#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess