Fit Simplified Poisson Log-Normal Model

poisson_lognormal uses maximum composite likelihood estimation to fit Poisson log-normal models to each sample.

poisson_lognormal_mcle(
  df_samples_subset,
  protein_names,
  condition,
  group,
  ncores = 1,
  slurm_settings = "slurm_batchtools.tmpl"
)

Arguments

df_samples_subset
Data frame or tibble with proteins counts, cell condition, and group information

protein_names
A vector of column names of protein to use in the analysis

condition
The column name of the condition variable

group
The column name of the group variable

ncores
Number of CPU cores

slurm_settings
Path to slurm cluster template for batchtools

Value

A list of class cytoeffect_poisson_mcle containing

tb_args

output tibble of model fits

protein_names

input protein names

condition

input condition variable

group

input group names

df_samples_subset

input df_samples_subset table

Examples

df = simulate_data(n_cells = 10)
str(df)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	80 obs. of  7 variables:
#>  $ donor    : chr  "pid01" "pid01" "pid01" "pid01" ...
#>  $ condition: Factor w/ 2 levels "control","treatment": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ m01      : num  16 39 14 15 44 61 34 2 13 30 ...
#>  $ m02      : num  12 8 3 1 6 20 18 2 30 10 ...
#>  $ m03      : num  3 9 5 0 8 4 4 4 8 2 ...
#>  $ m04      : num  2 0 0 0 1 8 1 2 10 6 ...
#>  $ m05      : num  6 29 4 1 22 2 5 5 11 2 ...

fit = poisson_lognormal_mcle(df,
                             protein_names = names(df)[3:ncol(df)],
                             condition = "condition",
                             group = "donor",
                             ncores = 1)

#> No readable configuration file found

#> Created registry in '/Users/cseiler/github_repos/cytoeffect/docs/reference/registry_2020_06_03_220549' using cluster functions 'Interactive'

#> Adding 8 jobs ...

#> Submitting 8 jobs in 8 chunks using cluster functions 'Multicore' ...

#> This deletes all files in '/Users/cseiler/github_repos/cytoeffect/docs/reference/registry_2020_06_03_220549'. Proceeding in 5 seconds ...

#> Unsetting registry as default

#> Recursively removing files in '/Users/cseiler/github_repos/cytoeffect/docs/reference/registry_2020_06_03_220549' ...